pyproject.toml
[project]
name = "ml-dl-api-dino"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "scikit-learn",
    "pillow",
    "python-multipart",
    "torch",
    "torchvision",
    "spacy"
]

[dependency-groups]
dev = ["pytest>=8.0", "httpx>=0.27"]
